Filter Field moved to separate file
